A few changes...
API Changes:
* Validate_checkAnn() now takes a PacketCrypt_ValidateCtx_t which is
* just a place to throw some scratch data while validating
* Validate_checkBlock() also takes the PacketCrypt_ValidateCtx_t
* Validate_checkBlock() now returns Validate_checkBlock_BAD_COINBASE if
* the coinbase commitment looks incorrect
* PacketCrypt_Coinbase_t layout changed, now must begin with the little
* endian word 0x0211f909
* BlockMiner_lockForMining() now outputs a structure BlockMiner_Hashes_t
* which contains a list of all announcement content hashes of
* announcements which were evicted from memory
* BlockMiner_addAnns() has a new parameter: noCopy, if true then the
* anns argument is included without copying, it is assumed that it can
* be passed to free() later

Hash computation changes:
* Validate_checkAnn: use 4 cycles instead of 3
* RandHash memory hardness: use the entire CryptoCycle_State_t when
* doing RandomHash rather than just a 32 byte segment of it
* RandGen: Change of the use of chacha20 prng because the way it was
* being used is different from use of chacha20 in the rest of the
* project
* Hash_compress() stop using a key when compressing because it makes the
* hash into a MAC and thus is incompatible with ordinary blake2b
* Hash_expand() stop passing the nonce number as the length to
* Hash_expand() because it makes it incompatible with ordinary chacha20
* CryptoCycle: move from 8 byte nonce to 12 byte nonce because this is
* what chacha20 normally wants
* CryptoCycle: add tzc (trailing zero count) in order to inform the
* algorithm how many bytes at the end of the message are not to be
* included in the authentication, needed for compatibility with ordinary
* chacha20poly1305 encryption
* CryptoCycle: Change of CryptoCycle_smul() because passing random
* values as a public key leads to assertion failure, now it takes a
* random value as a private key, computes the public key, and then
* computes the shared secret using a second random value as the other
* private key
* CryptoCycle: CryptoCycle_init() nolonger performs an encryption
* operation to save CPU cycles

Validate.c
* BUG: Validate_checkAnn() bump the number of cycles from 3 to 4 because
* previously there were only 3 cycles which is incorrect
* Validate_checkAnn() Put parenthecies around
* (CryptoCycle_getItemNo(&state) % Announce_TABLE_SZ); -- paranoia
* BUG: Validate_checkBlock() inverted an if:  if (effectiveAnnTarget >
* coinbaseCommitment->annLeastWorkTarget) {

RandHash.h
* Remove RandHash_Program_t and replace it with a simple buffer of fixed
* size
* Remove RandHash_generate() RandHash_compile() RandHash_freeProgram()
* which are not used
* Drop RandHash_ENOMEM which is never returned

RandHash.c -> deleted, was entirely unused cruft

RandHash_interepreted.c
* drop unused RandHash_execute_interpreted()
* superior debugging and testing macros for use with optest
* out4() use a macro to access the 4 ints for compatibility with
* OpTemplate2.h

RandGen.c
* Change of the use of chacha20 prng because the way it was being used
* is different from use of chacha20 in the rest of the project
* BUG: if the number of instructions went over the max, it would not
* report properly and RandHash_interpret would fail during execution.
* Better debugging
* Simplified API which doesn't require the caller to pass the max length
* which is consensus critical anyway

OpTemplate.h
* BUG: CLZ functions were returning wrong results because the result
* needs to be biased

Hash.h Hash.c
* Hash_compress() stop using a key when compressing because it makes the
* hash into a MAC and thus is incompatible with ordinary blake2b
* Hash_expand() stop passing the nonce number as the length to
* Hash_expand() because it makes it incompatible with ordinary chacha20

Difficulty.c
* BUG: Misuse of BN_copy() which should only be used with an
* uninitialized output

CryptoCycle.h
* Change of header layout, move from 8 byte nonce to 12 byte nonce
* because this is what chacha20 normally wants
* Change of header layout, add tzc (trailing zero count) in order to
* inform the algorithm how many bytes at the end of the message are not
* to be included in the authentication, needed for compatibility with
* ordinary chacha20poly1305 encryption
* Added progbuf to allow the RandHash program to be stored off of the
* stack because it is 8192 bytes

CryptoCycle.c
* Change of CryptoCycle_smul() because passing random values as a public
* key leads to assertion failure, now it takes a random value as a
* private key, computes the public key, and then computes the shared
* secret using a second random value as the other private key
* _update now takes an external progbuf rather than allocating it on the
* stack
* CryptoCycle_init() nolonger performs an encryption operation to save
* CPU cycles
* tzc and nonce size increase described in CryptoCycle.h

Conf.h
* Introduce Conf_PacketCrypt_ANN_FAKE_PARENT_HEIGHT which is used by
* BlockMiner to know how to handle announcements in the first 3 blocks

BlockMiner.c
* thread() Take out a lock every time we stop in order to be sure that
* the memory is synced
* waitState() Take out a lock when checking state of workers (to be sure
* that the memory is synced)
* waitState() wait 10 seconds instead of 1 to avoid spurious assertions
* on slower machines
* BlockMiner_free() take out a lock before asking the threads to stop
* BlockMiner_addAnns() implement noCopy to allow BlockMiner to take
* ownership of the announcement buffer
* BlockMiner_lockForMining() implement deletedAnnsOut to inform the
* caller of which announcement content is nolonger needed
* BlockMiner_lockForMining() allow announcements which have
* blockParentHeight of Conf_PacketCrypt_ANN_FAKE_PARENT_HEIGHT if the
* next block height is less than Conf_PacketCrypt_ANN_WAIT_PERIOD (to
* allow the first 3 blocks to be mined)

AnnMiner.c
* BUG: annHash() switch from 3 cycles to 4 because 3 cycles is incorrect
* annHash() parenths around (CryptoCycle_getItemNo(&w->state) %
* Announce_TABLE_SZ) because paranoia
* annHash() pass through progbuf to CryptoCycle because of change of API
* BUG: search() pass a newly allocated announcement instead of the
* memory location of the announcement buffer
* search() accept -1 return from write() and don't crash (it means the
* pipe was closed)
* BUG: mainLoop() when hardNonce reaches 0x7fffffff, stop the threads
* properly instead of just abandoning them

OpTemplate2.h
* Added a new file which should be exactly the same as OpTemplate.h
* except it is able to 'generate' golang code

optest.c
* Added a new conformance fuzz test which verifies that OpTemplate2.h
* behaves the same as OpTemplate.h

anngen.js - package.json
* Added a script to generate announcements with specific announcement
* content

+++ src/Hash.h
@@ -7,7 +7,7 @@
 
 void Hash_compress64(uint8_t output[static 64], uint8_t* buff, uint32_t len);
 void Hash_compress32(uint8_t output[static 32], uint8_t* buff, uint32_t len);
-void Hash_expand(uint8_t* buff, uint32_t len, const uint8_t seed[static 32], uint64_t num);
+void Hash_expand(uint8_t* buff, uint32_t len, const uint8_t seed[static 32], uint32_t num);
 void Hash_printHex(uint8_t* hash, int len);
 void Hash_eprintHex(uint8_t* hash, int len);
